+++ packages/load/src/load_package.ts
@@ -1,5 +1,3 @@
-import { Logger } from '@faasjs/logger'
-
 export type NodeRuntime = 'commonjs' | 'module'
 
 let _runtime: NodeRuntime | null = null
@@ -9,14 +7,14 @@ export function resetRuntime(): void {
 }
 
 /**
- * Detect the current JavaScript runtime environment.
+ * Detect current JavaScript runtime environment.
  *
- * This function checks for the presence of `import.meta` and `require` to determine
- * whether the runtime is using ECMAScript modules (ESM) or CommonJS modules (CJS).
+ * This function checks for presence of `import.meta` and `require` to determine
+ * whether runtime is using ECMAScript modules (ESM) or CommonJS modules (CJS).
  *
- * @returns {NodeRuntime} - Returns 'module' if the runtime is using ECMAScript modules,
+ * @returns {NodeRuntime} - Returns 'module' if runtime is using ECMAScript modules,
  *                            and 'cjs' if the runtime is using CommonJS modules.
- * @throws {Error} - Throws an error if the runtime cannot be determined.
+ * @throws {Error} - Throws an error if runtime cannot be determined.
  */
 export function detectNodeRuntime(): NodeRuntime {
   if (_runtime) return _runtime
@@ -29,18 +27,13 @@ export function detectNodeRuntime(): NodeRuntime {
   throw Error('Unknown runtime')
 }
 
-let tsxLoaded: null | boolean = null
-const logger = new Logger('loadPackage')
-
 /**
  * Asynchronously loads a package by its name, supporting both ESM and CommonJS runtimes.
  *
- * Also supports loading TypeScript and TSX files by checking for the presence of the "tsx" package.
- *
- * @template T - The type of the module to be loaded.
- * @param {string} name - The name of the package to load.
- * @returns {Promise<T>} A promise that resolves to the loaded module.
- * @throws {Error} If the runtime is unknown.
+ * @template T - The type of module to be loaded.
+ * @param {string} name - The name of package to load.
+ * @returns {Promise<T>} A promise that resolves to loaded module.
+ * @throws {Error} If runtime is unknown.
  *
  * @example
  * ```typescript
@@ -56,19 +49,6 @@ export async function loadPackage<T = unknown>(
   let module: any
 
   if (runtime === 'module') {
-    if (tsxLoaded === null) {
-      try {
-        // @ts-expect-error
-        await import('tsx')
-        tsxLoaded = true
-      } catch (_) {
-        tsxLoaded = false
-        logger.warn(
-          'Recommend installing the "tsx" package for loading ts/tsx files'
-        )
-      }
-    }
-
     module = await import(name)
 
     if (typeof defaultNames === 'string')
@@ -80,18 +60,6 @@ export async function loadPackage<T = unknown>(
   }
 
   if (runtime === 'commonjs') {
-    if (tsxLoaded === null) {
-      try {
-        globalThis.require('tsx')
-        tsxLoaded = true
-      } catch (_) {
-        tsxLoaded = false
-        logger.warn(
-          'Recommend installing the "tsx" package for loading ts/tsx files'
-        )
-      }
-    }
-
     module = globalThis.require(name)
 
     if (typeof defaultNames === 'string')