+++ packages/server/src/server/routes.ts
@@ -7,11 +7,8 @@ export function getRouteFiles(root: string, path: string): string[] {
 
   const searchPaths = [
     `${normalizedPath}.func.ts`,
-    `${normalizedPath}.func.tsx`,
     `${normalizedPath}${sep}index.func.ts`,
-    `${normalizedPath}${sep}index.func.tsx`,
     `${normalizedPath}${sep}default.func.ts`,
-    `${normalizedPath}${sep}default.func.tsx`,
   ]
 
   let currentPath = normalizedPath
@@ -23,10 +20,7 @@ export function getRouteFiles(root: string, path: string): string[] {
 
     if (currentPath.length < normalizedRoot.length - 1) break
 
-    searchPaths.push(
-      `${currentPath}${sep}default.func.ts`,
-      `${currentPath}${sep}default.func.tsx`
-    )
+    searchPaths.push(`${currentPath}${sep}default.func.ts`)
   }
 
   return searchPaths